+++ changelog.d/20240802_102901_rra_DM_45281.md
@@ -1,0 +1,3 @@
+### New features
+
+- Add new `safir.arq.WorkerSettings` class that models the acceptable parameters for an arq `WorkerSettings` object or class that Safir applications have needed.
